extend reusable server for testing
